feat: 이미 탈퇴한 회원에 대한 처리 추가

+++ src/main/java/com/tikitaka/triptroop/user/service/UserService.java
@@ -95,6 +95,10 @@ public class UserService {
     public void withdrawal(Long userId, String currentPassword) {
 
         final User user = validatePassword(userId, currentPassword);
+        
+        if (user.isWithdrawnUser()) {
+            throw new BadRequestException(ExceptionCode.ALREADY_WITHDRAWN_USER);
+        }
 
         userRepository.delete(user);
     }
